Refactor: CMS model 변경에 따른 ArticlePreview interface를 수정한다
